Move selectoptionsfromarray function logic to ContentHelperRuntime
